update validate clone plan

+++ lib/screens/order_screen/detail_order_screen.dart
@@ -693,7 +693,7 @@ class _OrderDetailScreenState extends State<OrderDetailScreen> {
                               product: ProductViewModel(
                                 id: detail.productId,
                                 name: detail.productName,
-                                price: detail.price!.toInt(),
+                                price: detail.price!,
                               ),
                               qty: detail.quantity));
                         }
